Make REPLs not to depend on their targets
Also remove the ‘repl_interpreted’ attribute which was an attribute I
introduced on my own and which I'm sure none is using. Further, the bug that
we're trying to fix is unavoidable when ‘repl_interpreted = False’, hence it
makes sense to always assume it to be ‘True’ satisfying the expectations
that users have for Haskell REPLs (as established by other tooling such as
Cabal and Stack).