Make REPL outputs force their dependencies
This fixes the regression introduced by #352: we have removed the dependency
on compiled “parent” target, but this also removed dependency on other
things that are necessary to run a REPL, such as:

* Haskell library dependencies
* pre-processed source files (hsc and chs)
* external shared libraries

And so they won't be built if a REPL is built.

This commit restores the dependencies allowing to run REPL without having to
build its “parent” target before that (still this does not cancel the
intended effects of #352).